@@ -1,29 +1,29 @@
 {
-  "_from": "@angular/animations@8.0.3",
-  "_id": "@angular/animations@8.0.3",
+  "_from": "@angular/animations@^8.2.14",
+  "_id": "@angular/animations@8.2.14",
   "_inBundle": false,
-  "_integrity": "sha512-9zciJ4YRR0bodFSYgsgXdYMz8wKKyVjch7XZADGkWubXT8mGuwlpdPMlQ6n9Cwj8Ebu0u52WxMeQsX76K9RlYA==",
+  "_integrity": "sha512-3Vc9TnNpKdtvKIXcWDFINSsnwgEMiDmLzjceWg1iYKwpeZGQahUXPoesLwQazBMmxJzQiA4HOMj0TTXKZ+Jzkg==",
   "_location": "/@angular/animations",
   "_phantomChildren": {},
   "_requested": {
-    "type": "version",
+    "type": "range",
     "registry": true,
-    "raw": "@angular/animations@8.0.3",
+    "raw": "@angular/animations@^8.2.14",
     "name": "@angular/animations",
     "escapedName": "@angular%2fanimations",
     "scope": "@angular",
-    "rawSpec": "8.0.3",
+    "rawSpec": "^8.2.14",
     "saveSpec": null,
-    "fetchSpec": "8.0.3"
+    "fetchSpec": "^8.2.14"
   },
   "_requiredBy": [
     "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/@angular/animations/-/animations-8.0.3.tgz",
-  "_shasum": "9fc82919b8eb28d4d1a9ff93095907acb2a64e91",
-  "_spec": "@angular/animations@8.0.3",
-  "_where": "C:\\Users\\adora\\OneDrive\\Desktop\\Angular8-SpringBoot-CRUD-Tutorial\\angular8-springboot-client",
+  "_resolved": "https://registry.npmjs.org/@angular/animations/-/animations-8.2.14.tgz",
+  "_shasum": "76736b21e56165e6ca4925fb69605bdcc56aba7d",
+  "_spec": "@angular/animations@^8.2.14",
+  "_where": "E:\\Abdul_Alim\\Angular8CRUD\\2nd CRUD",
   "author": {
     "name": "angular"
   },
@@ -68,7 +68,10 @@
     ]
   },
   "peerDependencies": {
-    "@angular/core": "8.0.3"
+    "@angular/core": "8.2.14"
+  },
+  "publishConfig": {
+    "registry": "https://wombat-dressing-room.appspot.com"
   },
   "repository": {
     "type": "git",
@@ -76,5 +79,5 @@
   },
   "sideEffects": false,
   "typings": "./animations.d.ts",
-  "version": "8.0.3"
+  "version": "8.2.14"
 }
