@@ -40,52 +39,50 @@
   "deprecated": false,
   "description": "The most popular front-end framework for developing responsive, mobile first projects on the web.",
   "devDependencies": {
-    "@babel/cli": "^7.7.4",
-    "@babel/core": "^7.7.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.7.4",
-    "@babel/preset-env": "^7.7.4",
-    "autoprefixer": "^9.7.2",
-    "babel-eslint": "^10.0.3",
-    "babel-plugin-istanbul": "^5.2.0",
-    "broken-link-checker": "^0.7.8",
-    "bundlesize": "^0.18.0",
+    "@babel/cli": "^7.11.6",
+    "@babel/core": "^7.11.6",
+    "@babel/preset-env": "^7.11.5",
+    "@rollup/plugin-babel": "^5.2.1",
+    "@rollup/plugin-commonjs": "^15.1.0",
+    "@rollup/plugin-node-resolve": "^9.0.0",
+    "autoprefixer": "^9.8.6",
+    "babel-plugin-istanbul": "^6.0.0",
+    "bundlewatch": "^0.3.1",
     "clean-css-cli": "^4.3.0",
-    "cross-env": "^5.2.1",
-    "eslint": "^5.16.0",
-    "find-unused-sass-variables": "^0.6.0",
+    "cross-env": "^7.0.2",
+    "eslint": "^7.11.0",
+    "eslint-config-xo": "^0.32.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-unicorn": "^22.0.0",
+    "find-unused-sass-variables": "^2.0.0",
     "glob": "^7.1.6",
     "hammer-simulator": "0.0.1",
-    "http-server": "^0.11.1",
     "ip": "^1.1.5",
-    "jquery": "^3.4.1",
-    "karma": "^3.1.4",
+    "jquery": "^3.5.1",
+    "karma": "^5.2.3",
     "karma-browserstack-launcher": "1.4.0",
     "karma-chrome-launcher": "^3.1.0",
-    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-detect-browsers": "^2.3.3",
-    "karma-firefox-launcher": "^1.2.0",
-    "karma-qunit": "^3.1.3",
+    "karma-firefox-launcher": "^1.3.0",
+    "karma-qunit": "^4.1.1",
     "karma-sinon": "^1.0.5",
-    "node-sass": "^4.13.0",
-    "nodemon": "^1.19.4",
+    "linkinator": "^2.1.3",
+    "lockfile-lint": "^4.3.7",
+    "node-sass": "^4.14.1",
+    "nodemon": "^2.0.4",
     "npm-run-all": "^4.1.5",
-    "popper.js": "^1.16.0",
-    "postcss-cli": "^6.1.3",
-    "qunit": "2.9.2",
-    "rollup": "1.26.5",
-    "rollup-plugin-babel": "^4.3.3",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "shelljs": "^0.8.3",
+    "popper.js": "^1.16.1",
+    "postcss-cli": "^7.1.2",
+    "qunit": "^2.11.3",
+    "rollup": "^2.29.0",
+    "shelljs": "^0.8.4",
     "shx": "^0.3.2",
     "sinon": "^7.5.0",
-    "stylelint": "^9.10.1",
-    "stylelint-config-twbs-bootstrap": "^0.5.0",
-    "uglify-js": "^3.6.9",
-    "vnu-jar": "19.9.4"
-  },
-  "engines": {
-    "node": ">=6"
+    "stylelint": "^13.7.2",
+    "stylelint-config-twbs-bootstrap": "^2.1.0",
+    "terser": "^4.8.0",
+    "vnu-jar": "20.6.30"
   },
   "files": [
     "dist/{css,js}/*.{css,js,map}",
