@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 @Injectable({
   providedIn: 'root'
@@ -11,23 +12,20 @@ export class EmployeeService {
 
   constructor(private http: HttpClient) { }
 
-  getEmployee(id: number): Observable<any> {
-    return this.http.get('http://localhost:8081/show-ms-user-by/' + id);
-  }
-
   createEmployee(employee: Object): Observable<Object> {
-    return this.http.post('http://localhost:8081/save-ms-user', employee);
+    return this.http.post(`${this.baseUrl}` + '/save-ms-user', employee).pipe(map(data => data));
   }
 
-  updateEmployee(id: number, value: any): Observable<Object> {
-    return this.http.put('http://localhost:8081/update-ms-user/' + id, value);
+  updateEmployee(value: any): Observable<Object> {
+    return this.http.post(`${this.baseUrl}` + '/update-ms-user', value).pipe(map(data => data));
   }
 
   deleteEmployee(id: number): Observable<any> {
-    return this.http.delete('http://localhost:8081/delete-ms-user/' + id, { responseType: 'text' });
+    return this.http.delete(`${this.baseUrl}` + '/delete-ms-user/' + id).pipe(map(data => data));
   }
 
   getEmployeesList(): Observable<any> {
-    return this.http.get('http://localhost:8081/show-ms-user');
+    return this.http.get(`${this.baseUrl}` + '/show-ms-user');
   }
+
 }
