@@ -1,16 +1,11 @@
 package com.exam.controller;
 
 import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-
-import javax.servlet.http.HttpServletRequest;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -19,16 +14,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.ModelAndView;
 
-import com.exam.commerz.Utility.ParameterBuilder;
-import com.exam.model.ApplyLoan;
-import com.exam.model.Contact;
-import com.exam.model.Installment;
+import com.exam.model.Response;
 import com.exam.model.UserInfo;
-import com.exam.service.ApplyLoanServiceImpl;
 import com.exam.service.ContactServiceImpl;
-import com.exam.service.InstallmentServiceImpl;
 import com.exam.service.UserInfoService;
 import com.exam.service.UserInfoServiceImpl;
 
@@ -47,12 +36,25 @@ public class Angular8RestController {
 	@Autowired
 	UserInfoService userInfoService;
 	
+
+	@Autowired
+	UserInfoServiceImpl userInfoServiceImpl;
 	
 	
 	@PostMapping(value = "/save-ms-user")
-	public String save(@RequestBody UserInfo entity) {
+	public Response saveEmp(@RequestBody UserInfo entity) {
 		
-		System.out.println("Save UserInfo::::::::::"+entity);
+		String pass = entity.getPassword();
+		entity.setPassword(passwordEncoder.encode(pass));
+		entity.setRole("User");
+		entity.setCreatedDate(new Date());
+		
+		return userInfoServiceImpl.saveEmp(entity);
+		
+	}
+	
+	@PostMapping(value = "/save-user")
+	public String save(@RequestBody UserInfo entity) {
 		
 		String pass = entity.getPassword();
 		entity.setPassword(passwordEncoder.encode(pass));
@@ -64,17 +66,17 @@ public class Angular8RestController {
 		
 	}
 	
-	@GetMapping("/delete-ms-user/{id}")
-	public List<UserInfo> deleteUserInfo(@PathVariable long id) {
+	@DeleteMapping("/delete-ms-user/{id}")
+	public Response deleteUserInfo(@PathVariable long id) {
 		
 		 userInfoService.delete(id);
-		 List<UserInfo> user = userInfoService.getAll();
-		 return user;
+		 Response rs = new Response();
+		 rs.setMessage("Delete Successfull !");
+		 return rs;
 	}
 
 	@GetMapping("/show-ms-user-by-username/{userName}")
 	public List<UserInfo> getUserByUsername(@PathVariable String userName){
-		System.out.println(userName+"::::::::::::::::::::::::::::::::::::::::::::::::::::::");
 		 List<UserInfo> entity = userInfoService.getByUsername(userName);
 		
 		return entity;
@@ -82,7 +84,6 @@ public class Angular8RestController {
 	
 	@GetMapping("/show-ms-user-by/{id}")
 	public UserInfo getUserById(@PathVariable("id") long id){
-		System.out.println(id+"::::::::::::::::::::::::::::::::::::::::::::::::::::::");
 		 UserInfo entity = userInfoService.getById(id);
 		
 		return entity;
@@ -92,24 +93,23 @@ public class Angular8RestController {
 	public List<UserInfo> showUserInfo(){
 		
 		List<UserInfo> user = userInfoService.getAll();
-		System.out.println(user+":::::::::::::::::::::::::::::::::::::::::::::::::::::");
 		return user;
 		
 	}
 	
-	@PostMapping(value = "/update-ms-user/{id}")
-	public String updateUser(@RequestBody UserInfo entity, @PathVariable("id") long id) {
-		
-		System.out.println("Save UserInfo::::::::::"+entity);
-		
+	@PostMapping(value = "/update-ms-user")
+	public Response updateUser(@RequestBody UserInfo entity) {
 			
-		UserInfo en = userInfoService.getById(id);
+		UserInfo en = userInfoService.getById(entity.getUserId());
+		Response rs = new Response();
+		if(en != null) {
+			entity.setUpdateDate(new Date());
+			entity.setFirstName(entity.getFirstName());
+			userInfoService.update(entity);
+			rs.setMessage("Hi "+ en.getUsername() +" Your's update is completed.");
+		}
 		
-		en.setUpdateDate(new Date());
-		en.setFirstName(entity.getFirstName());
-		userInfoService.update(en);
-		
-		return "Hi "+ en.getUsername() +" Your's update is completed.";
+		return rs;
 		
 	}
 
